draft: establish canonical posetag package and CLI shims
