+++ src/gt6dof_atag/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility CLI entry points for legacy gtat commands."""
