+++ src/gt6dof_atag/cli/annotate.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-annotate command."""
+
+from posetag.cli.annotate import main
