+++ src/gt6dof_atag/cli/capture_face.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-capture-face command."""
+
+from posetag.cli.capture_face import main
