+++ src/gt6dof_atag/cli/charuco.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-calib-charuco command."""
+
+from posetag.cli.charuco import main
