+++ src/gt6dof_atag/cli/collect.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-collect command."""
+
+from posetag.cli.collect import main
