+++ src/gt6dof_atag/cli/gen_tags.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-gen-tags command."""
+
+from posetag.cli.gen_tags import main
