+++ src/gt6dof_atag/cli/make_board.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the legacy gtat-make-board command."""
+
+from posetag.cli.make_board import main
