+++ src/gt6dof_atag/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility wrappers for legacy gt6dof_atag utility imports."""
