+++ src/gt6dof_atag/utils/logger.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for legacy gt6dof_atag logger imports."""
+
+from posetag.utils.logger import *  # noqa: F401,F403
