+++ src/gt6dof_atag/utils/project_config.py
@@ -1,0 +1,4 @@
+"""Compatibility wrapper for legacy gt6dof_atag project configuration imports."""
+
+from posetag.utils.project_config import *  # noqa: F401,F403
+from posetag.utils.project_config import _norm  # noqa: F401
