+++ src/gt6dof_atag/utils/project_paths.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for legacy gt6dof_atag path helper imports."""
+
+from posetag.utils.project_paths import *  # noqa: F401,F403
