+++ src/posetag/__init__.py
@@ -1,0 +1,3 @@
+"""Canonical PoseTag package namespace."""
+
+__version__ = "0.1.0"
