+++ src/posetag/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Canonical PoseTag CLI entry points."""
