+++ src/posetag/cli/annotate.py
@@ -1,0 +1,7 @@
+"""PoseTag CLI wrapper for per-face annotation."""
+
+
+def main():
+    from annotate_shots import main as _main
+
+    return _main()
