+++ src/posetag/cli/capture_face.py
@@ -1,0 +1,7 @@
+"""PoseTag CLI wrapper for face-shot capture."""
+
+
+def main():
+    from capture_face import main as _main
+
+    return _main()
