+++ src/posetag/cli/collect.py
@@ -1,0 +1,7 @@
+"""PoseTag CLI wrapper for dataset collection."""
+
+
+def main():
+    from collect_gt_dataset import main as _main
+
+    return _main()
