+++ src/posetag/cli/gen_tags.py
@@ -1,0 +1,7 @@
+"""PoseTag CLI wrapper for AprilTag sheet generation."""
+
+
+def main():
+    from gen_april_tags import main as _main
+
+    return _main()
