+++ src/posetag/cli/make_board.py
@@ -1,0 +1,7 @@
+"""PoseTag CLI wrapper for AprilTag board creation."""
+
+
+def main():
+    from make_board import main as _main
+
+    return _main()
