+++ src/posetag/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Canonical PoseTag utility import surface."""
