+++ src/posetag/utils/logger.py
@@ -1,0 +1,3 @@
+"""Canonical PoseTag wrapper for logging helpers."""
+
+from utils.logger import *  # noqa: F401,F403
