+++ src/posetag/utils/project_config.py
@@ -1,0 +1,6 @@
+"""Canonical PoseTag wrapper for project configuration helpers."""
+
+from utils import project_config as _project_config
+from utils.project_config import *  # noqa: F401,F403
+
+_norm = _project_config._norm
