+++ src/posetag/utils/project_paths.py
@@ -1,0 +1,3 @@
+"""Canonical PoseTag wrapper for project path helpers."""
+
+from utils.project_paths import *  # noqa: F401,F403
